# backend/openapi/openapi.yaml
components:
    schemas:
...
        httpx.ErrorBody:
            properties:
                error:
                    $ref: '#/components/schemas/httpx.ErrorDetail'
            type: object
        httpx.ErrorDetail:
            properties:
                code:
                    type: string
                fields:
                    additionalProperties:
                        type: string
                    type: object
                message:
                    type: string
                request_id:
                    type: string
            type: object
        identity.OTPRequested:
            properties:
                dev_code:
                    type: string
                expires_at:
                    format: date-time
                    type: string
            type: object
        identity.PhoneRequest:
            properties:
                phone:
                    type: string
            required:
                - phone
            type: object
        identity.RefreshRequest:
            properties:
                refresh_token:
                    type: string
            required:
                - refresh_token
            type: object
        identity.Session:
            properties:
                access_token:
                    type: string
                expires_at:
                    format: date-time
                    type: string
                refresh_expires_at:
                    format: date-time
                    type: string
                refresh_token:
                    type: string
                token_type:
                    type: string
                user:
                    $ref: '#/components/schemas/identity.User'
            type: object
        identity.Tokens:
            properties:
                access_token:
                    type: string
                expires_at:
                    format: date-time
                    type: string
                refresh_expires_at:
                    format: date-time
                    type: string
                refresh_token:
                    type: string
                token_type:
                    type: string
            type: object
        identity.UpdateMeRequest:
            properties:
                district:
                    nullable: true
                    type: string
                language:
                    nullable: true
                    type: string
                name:
                    nullable: true
                    type: string
            type: object
        identity.User:
            properties:
                created_at:
                    format: date-time
                    type: string
                district:
                    type: string
                id:
                    type: string
                language:
                    type: string
                name:
                    type: string
                phone:
                    type: string
                role:
                    type: string
            type: object
        identity.VerifyRequest:
            properties:
                code:
                    type: string
                phone:
                    type: string
            required:
                - code
                - phone
            type: object
...
    securitySchemes:
        bearer:
            bearerFormat: JWT
            scheme: bearer
            type: http
info:
    description: Offline-first farming assistant API. Error `message` values are i18n dictionary keys.
    title: AgriSmart API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /v1/auth/guest:
        post:
            operationId: post_v1_auth_guest
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.Session'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            summary: Start a guest session
            tags:
                - auth
    /v1/auth/logout:
        post:
            operationId: post_v1_auth_logout
            responses:
                "204":
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            security:
                - bearer: []
            summary: Revoke the current session
            tags:
                - auth
            x-permission: profile:read
    /v1/auth/otp/request:
        post:
            operationId: post_v1_auth_otp_request
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/identity.PhoneRequest'
                required: true
            responses:
                "202":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.OTPRequested'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            summary: Send a one-time code by SMS
            tags:
                - auth
    /v1/auth/otp/verify:
        post:
            operationId: post_v1_auth_otp_verify
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/identity.VerifyRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.Session'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            summary: Exchange a code for a session
            tags:
                - auth
    /v1/auth/refresh:
        post:
            operationId: post_v1_auth_refresh
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/identity.RefreshRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.Tokens'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            summary: Rotate a refresh token
            tags:
                - auth
...
    /v1/me:
        get:
            operationId: get_v1_me
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.User'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            security:
                - bearer: []
            summary: Current profile
            tags:
                - profile
            x-permission: profile:read
        patch:
            operationId: patch_v1_me
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/identity.UpdateMeRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/identity.User'
                    description: success
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/httpx.ErrorBody'
                    description: error
            security:
                - bearer: []
            summary: Update profile
            tags:
                - profile
            x-permission: profile:write
...